+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # create-llama
 
+## 0.5.3
+
+### Patch Changes
+
+- 76ec360: Update templates to use new chat ui config
+
+## 0.5.2
+
+### Patch Changes
+
+- c9f8f8d: Use custom component for deep research use case
+
 ## 0.5.1
 
 ### Patch Changes